AMPOPS-230 AMP 2.4.5->2.4.6 merged into AMP 2.5
git-svn-id: file:///svn/branches/AMP_2_5_RELEASE@20719 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/dgfoundation/amp/ar/AmpReportGenerator.java
@@ -66,8 +66,9 @@ public class AmpReportGenerator extends ReportGenerator {
 	List<AmpReportColumn> extractable; // columns extractable while building the report
 	protected int extractableCount; // looks like extractable.size() [need to check!]
 	private List<String> columnsToBeRemoved;
-	private boolean debugMode=false;
-	private boolean pledgereport=false;
+	private boolean debugMode = false;
+	private boolean pledgereport = false;
+	private boolean cleanupMetadata = true;
 	
 	private HttpSession session	= null;
 	private BigDecimal totalac;
@@ -882,7 +883,8 @@ public class AmpReportGenerator extends ReportGenerator {
 		removeUnrenderizableData(report); // apply year range filter from settings and any other "non-displaying" filtering which SHOULD NOT affect the report except rendering
 		
 		report.getAllCells(listOfCells, true); //repeatedly fetch cells, as some might have been added in the meantime (postprocessing)
-		deleteMetadata(listOfCells);
+		if (getCleanupMetadata())
+			deleteMetadata(listOfCells);
 		System.out.format("AmpReportGenerator: AmountCell.getPercentage calls = %d, iterations = %d, iterations / call = %.2f\n", AmountCell.getPercentageCalls, AmountCell.getPercentageIterations, 1.0 * AmountCell.getPercentageIterations / (0.01 + AmountCell.getPercentageCalls));
 		System.out.format("AmpReportGenerator: AmountCell.getAmountWithMergedCells calls = %d, iterations = %d, iterations / call = %.2f\n", AmountCell.merged_cells_get_amount_calls, AmountCell.merged_cells_get_amount_iterations, 1.0 * AmountCell.merged_cells_get_amount_iterations / (0.01 + AmountCell.merged_cells_get_amount_calls));
 	}
@@ -929,6 +931,20 @@ public class AmpReportGenerator extends ReportGenerator {
 		}
 	}	
 	
+	/** somewhat stupid heuristic: the first number in the "year" name is a year and we can range on that */
+	public final static Integer getYearInteger(String yearStr)
+	{
+		java.util.StringTokenizer tok = new java.util.StringTokenizer(yearStr);
+		while (tok.hasMoreTokens())
+		{
+			String z = tok.nextToken();
+			Integer k = getInteger(z);
+			if (k != null)
+				return k;
+		}
+		return null;
+	}
+	
 	protected void applyYearRangeSettings(GroupColumn column)
 	{
 		Iterator it = column.iterator();
@@ -938,7 +954,7 @@ public class AmpReportGenerator extends ReportGenerator {
 			if (item instanceof Column)
 			{
 				String yearStr = ((Column) item).getName();
-				Integer year = getInteger(yearStr);
+				Integer year = getYearInteger(yearStr);
 				if (year != null)
 					if (!filter.passesYearRangeFilter(year))
 						it.remove();
@@ -1220,4 +1236,13 @@ public class AmpReportGenerator extends ReportGenerator {
 		return fakeC;
 	}
 
+	public boolean getCleanupMetadata()
+	{
+		return cleanupMetadata;
+	}
+	
+	public void setCleanupMetadata(boolean cleanupMetadata)
+	{
+		this.cleanupMetadata = cleanupMetadata;
+	}
 }
